requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
coverage==7.3.0
ecdsa==0.18.0
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.0
pytest==7.4.1
pytest-cov==4.1.0
pytest-flask-sqlalchemy==1.1.0
pytest-mock==3.11.1
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.7
